backend/requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
pytesseract
numpy
opencv-python-headless
pillow
python-dotenv
pydantic[all]
